+++ internal/tools/tools.py
@@ -1,0 +1,7 @@
+from internal.tools.basic_tools import add_basic_tools
+
+
+def add_all_tools(agent) -> None:
+    """Add tools to the agent."""
+
+    add_basic_tools(agent)
